Feat: biasScore 데이터 타입 Int에서 Double로 변경
Former-commit-id: 1f229f86b03734aaaf9c05165fc1580870c2374f

+++ android/app/src/main/java/com/ssafy/firstproject/data/model/NewsArticle.kt
@@ -1,7 +1,7 @@
 package com.ssafy.firstproject.data.model
 
 data class NewsArticle(
-    val biasScore: Int,
+    val biasScore: Double,
     val content: String,
     val description: String,
     val extractedAt: List<Int>,
